Add a laser.

+++ src/main/java/com/netchip/LaserMachines/api/ILaserTile.java
@@ -2,5 +2,5 @@ package com.netchip.LaserMachines.api;
 
 public interface ILaserTile {
     public boolean needsPower();
-    public float transferPower(int amountPower);
+    public int transferPower(int amountPower);
 }
